requirements.txt
asgiref==3.2.3
Django==3.0.4
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-jsonapi==3.1.0
drf-nested-routers==0.91
inflection==0.3.1
pytz==2018.4
sqlparse==0.3.1
virtualenv==16.0.0
virtualenvwrapper-win==1.2.5
